Use robust method of checking provisioning subnet
Use ipcalc for checking that the provisioning interface is a member of
the correct network. This replaces some grep-based shenanigans that made
unjustified assumptions about the format of IPv6 addresses.

It also corrects the behaviour on IPv4 to match IPv6, so that we check
for IPs on the same subnet as the provisioning VIP, not for the exact
provisioning VIP to appear. This should prevent a period of crashlooping
when the metal3 pod moves to a new host.

Also update the error message to reflect the actual semantics, which
are: if the provisioning interface has any IP addresses, one of them
must be for a subnet that also contains the provisioning VIP.

It appears that the purpose of the test is to prevent the provisioning
network being accidentally enabled on a machine network interface, and
reconfiguring that interface. When the provisioning network is managed,
an IP in the provisioning subnet should only appear when this check has
previously succeeded on some other host that was previously running
dnsmasq, so it should be safe to ignore the presence of other IPs. When
the provisioning network is unmanaged it is fine to have other IPs, and
the presence of the provisioning subnet indicates that we have chosen
the correct interface.

(cherry picked from commit db23d8d3226cb3b66686d064df5bd3c7f6f70197)